fix import slack theme test
fix the test so it doesn't depend on the default theme values, AND so that it tests using values unlikely to already be used.

From 5.29 to 5.34 the slack theme got 2 additional colors at indices 8 and 9. These are unused by MM but their presence is verified, so I've adjusted the slackThemeProps to add them w/ names that will not be added to the MM theme.
fix import slack theme test

fix the test so it doesn't depend on the default theme values, AND so that it tests using values unlikely to already be used.

From 5.29 to 5.34 the slack theme got 2 additional colors at indices 8 and 9. These are unused by MM but their presence is verified, so I've adjusted the slackThemeProps to use null as their MM prop names so they are not added to the expected new MM theme.